get_cred: relax ticket check
and allow the realm to be canonicalized. This is
useful when Windows KDC returns the canonical realm
instead of requested lc or netbios realm.

TODO: Is this ok ? in find_cred we seem to be okay
with any realm.
The old code didn't check at all it seems (check
when this was introduced).

Signed-off-by: Isaac Boukris <iboukris@gmail.com>